Solve timing bugs. Closes #13, #14.
We handle fragmentation by sending an invalid packet to the server after
every "command" packet. After handling our command, the server dutifully
sends us an error message about the invalid packet. When we receive this
error, we know we have the complete response from the server.

Core gets a rewrite because an entire class just to wrap a socket seems
excessive to my older, wiser eyes.